+++ packages/twenty-sdk/scripts/remote-dom/generateRemoteDomElements.ts
@@ -3,11 +3,11 @@ import * as fs from 'fs';
 import * as path from 'path';
 import { IndentationText, Project, QuoteKind } from 'ts-morph';
 
-import { ALLOWED_HTML_ELEMENTS } from '../../src/front-component/common/AllowedHtmlElements';
-import { ALLOWED_UI_COMPONENTS } from '../../src/front-component/common/AllowedUiComponents';
-import { COMMON_HTML_EVENTS } from '../../src/front-component/common/CommonHtmlEvents';
-import { EVENT_TO_REACT } from '../../src/front-component/common/EventToReact';
-import { HTML_COMMON_PROPERTIES } from '../../src/front-component/common/HtmlCommonProperties';
+import { ALLOWED_HTML_ELEMENTS } from '../../src/sdk/front-component-common/AllowedHtmlElements';
+import { ALLOWED_UI_COMPONENTS } from '../../src/sdk/front-component-common/AllowedUiComponents';
+import { COMMON_HTML_EVENTS } from '../../src/sdk/front-component-common/CommonHtmlEvents';
+import { EVENT_TO_REACT } from '../../src/sdk/front-component-common/EventToReact';
+import { HTML_COMMON_PROPERTIES } from '../../src/sdk/front-component-common/HtmlCommonProperties';
 
 import {
   type ComponentSchema,
