+++ packages/twenty-sdk/src/front-component/remote/components/FrontComponentErrorEffect.tsx
