+++ packages/twenty-sdk/src/front-component/remote/components/FrontComponentUpdateContextEffect.tsx
@@ -1,5 +1,5 @@
-import { type FrontComponentExecutionContext } from '@/front-component/renderer/types/FrontComponentExecutionContext';
-import { type WorkerExports } from '@/front-component/renderer/types/WorkerExports';
+import { type FrontComponentExecutionContext } from '@/front-component/types/FrontComponentExecutionContext';
+import { type WorkerExports } from '@/front-component/types/WorkerExports';
 import { type ThreadWebWorker } from '@quilted/threads';
 import { useEffect } from 'react';
