+++ packages/twenty-sdk/src/front-component/remote/components/FrontComponentWorkerEffect.tsx
