+++ packages/twenty-sdk/src/front-component/remote/mock/mock-front-component.ts
