+++ packages/twenty-sdk/src/front-component/remote/worker/remote-worker.ts
