+++ packages/twenty-sdk/src/front-component/types/PropertySchema.ts
