+++ packages/twenty-sdk/src/sdk/front-component-api/types/FrontComponentExecutionContextStore.ts
