+++ packages/twenty-sdk/src/sdk/front-component-common/CommonHtmlEvents.ts
