+++ packages/twenty-sdk/src/sdk/front-component-common/HtmlCommonProperties.ts
@@ -1,4 +1,4 @@
-import { type PropertySchema } from '../renderer/types/PropertySchema';
+import { type PropertySchema } from '@/front-component/types/PropertySchema';
 
 export const HTML_COMMON_PROPERTIES: Record<string, PropertySchema> = {
   id: { type: 'string', optional: true },
